Split layout.lua into smaller modules (#168)
